% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSVG.R
\name{convertSVG}
\alias{convertSVG}
\title{convertSVG}
\usage{
convertSVG(svg, device = "png", width = 8.2677, height = 11.6929,
  dpi = 300)
}
\arguments{
\item{svg}{svg file}

\item{device}{target format}

\item{width}{output width}

\item{height}{output height}

\item{dpi}{output dpi}
}
\value{
invisible grob object
}
\description{
convert svg to png or other format
}
\author{
Zhaodong Hao, Dekang Lv, Guangchuang Yu, Ying Ge, Jisen Shi, Jinhui Chen
}
