% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFFex.R
\name{GFFex}
\alias{GFFex}
\title{GFFex}
\usage{
GFFex(input, karyotype, feature = "gene", window = 1000000)
}
\arguments{
\item{input}{gff file}

\item{karyotype}{karyotype file}

\item{feature}{feature format}

\item{window}{window size}
}
\value{
dataframe
}
\description{
extract some specific feature information from a gff file
}
\author{
Zhaodong Hao, Dekang Lv, Ying Ge, Jisen Shi, Dolf Weijers, Guangchuang Yu, Jinhui Chen
}
