% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideogram.R
\name{ideogram}
\alias{ideogram}
\title{ideogram}
\usage{
ideogram(karyotype, overlaid, label = NULL, colorset, width, Lx, Ly,
  output = "chromosome.svg")
}
\arguments{
\item{karyotype}{karyotype data}

\item{overlaid}{overlaid data}

\item{label}{track label data}

\item{colorset}{overlaid heatmap color}

\item{width}{width of plot region}

\item{Lx}{position of legend (x)}

\item{Ly}{position of legend (y)}

\item{output}{output file, only svg is supported}
}
\value{
output file
}
\description{
ideogram with overlaid heatmap annotation and optional track label
}
\examples{
# Loading the package
require(RIdeogram)

# Loading the testing data
data(karyotype, package="RIdeogram")
data(mydata, package="RIdeogram")
data(mydata_interval, package="RIdeogram")

# Checking the data format
head(karyotype)
head(mydata)
head(mydata_interval)

# Running the function
svgfile <- tempfile(fileext = ".svg")
pdffile <- tempfile(fileext = ".pdf")
ideogram(karyotype, mydata, mydata_interval, c("navy", "white", "firebrick3"), 170, 160, 35, svgfile)
svg2pdf(svgfile, pdffile)

# Then, you will find a PDF file in your Working Directory.

}
\author{
Zhaodong Hao, Dekang Lv, Guangchuang Yu, Ying Ge, Jisen Shi, Jinhui Chen
}
