\name{liriodendron_karyotype}
\docType{data}
\alias{liriodendron_karyotype}
\title{Karyotype information of the Liriodendron genome}
\description{
  Liriodendron chinense genome
}
\usage{data(liriodendron_karyotype)}
\format{data frame}
\source{Nature Plants, doi: 10.1038/s41477-018-0323-6}
\keyword{datasets}
