\name{Pi_for_CE}
\docType{data}
\alias{Pi_for_CE}
\title{Pi of one Liriodendron group}
\description{
  Pi of the China east group of Liriodendron chinense and being calculated in a 2-Mb sliding window with a
1-Mb step
}
\usage{data(Pi_for_CE)}
\format{data frame}
\source{Nature Plants, doi: 10.1038/s41477-018-0323-6}
\keyword{datasets}
