\name{Pi_for_CE_and_CW}
\docType{data}
\alias{Pi_for_CE_and_CW}
\title{Pi of two Liriodendron groups}
\description{
  Pi values of the China east and west groups of Liriodendron chinense and being calculated in a 2-Mb sliding window with a
1-Mb step
}
\usage{data(Pi_for_CE_and_CW)}
\format{data frame}
\source{Nature Plants, doi: 10.1038/s41477-018-0323-6}
\keyword{datasets}
