\name{Random_RNAs_500}
\docType{data}
\alias{Random_RNAs_500}
\title{500 RNAs' position}
\description{
  500 RNAs randomly selected from all tRNAs, rRNAs and miRNA in the human genome.
}
\usage{data(Random_RNAs_500)}
\format{data frame}
\source{Gencode (https://www.gencodegenes.org/)}
\keyword{datasets}
