% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg2jpg.R
\name{svg2jpg}
\alias{svg2jpg}
\title{svg2jpg}
\usage{
svg2jpg(svg, jpg, width = 210, height = 297, dpi = 300)
}
\arguments{
\item{svg}{svg file}

\item{jpg}{jpg file}

\item{width}{output width}

\item{height}{output height}

\item{dpi}{output dpi}
}
\value{
invisible grob object
}
\description{
convert svg to jpg
}
\author{
Zhaodong Hao, Dekang Lv, Guangchuang Yu, Ying Ge, Jisen Shi, Jinhui Chen
}
