% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg2pdf.R
\name{svg2pdf}
\alias{svg2pdf}
\title{svg2pdf}
\usage{
svg2pdf(svg, pdf, width = 210, height = 297)
}
\arguments{
\item{svg}{svg file}

\item{pdf}{pdf file}

\item{width}{output width}

\item{height}{output height}
}
\value{
invisible grob object
}
\description{
convert svg to pdf
}
\author{
Guangchuang Yu
}
