\name{synteny_ternary_comparison_graident}
\docType{data}
\alias{synteny_ternary_comparison_graident}
\title{Synteny for three genome comparison with gradient fill}
\description{
  Genome Synteny among Amborella, Grape and Liriodendron with gradient fill
}
\usage{data(synteny_ternary_comparison_graident)}
\format{data frame}
\source{MCscan}
\keyword{datasets}
