\name{Fst_between_CE_and_CW}
\docType{data}
\alias{Fst_between_CE_and_CW}
\title{Fst between two Liriodendron groups}
\description{
  Fst between China east and west groups of Liriodendron chinense and is calculated in a 2-Mb sliding window with a
1-Mb step
}
\usage{data(Fst_between_CE_and_CW)}
\format{data frame}
\source{Nature Plants, doi: 10.1038/s41477-018-0323-6}
\keyword{datasets}
